+++ crates/orderbook/src/api/get_fee_info.rs
@@ -1,7 +1,8 @@
 use crate::{
-    api::{convert_json_response, order_validation::PartialValidationError, IntoWarpReply},
+    api::{convert_json_response, IntoWarpReply},
     fee::FeeData,
     fee::MinFeeCalculating,
+    order_validation::PartialValidationError,
 };
 use anyhow::Result;
 use chrono::{DateTime, Utc};
